feat(BO-1808): retry persistent subscriptions if they fail
