+++ build.gradle.kts
@@ -25,7 +25,6 @@ repositories {
 dependencies {
     val jacksonVersion = "2.11.3"
     val testContainersVersion = "1.16.2"
-    val micrometerVersion = "1.7.0"
 
     api("com.github.msemys:esjc:2.4.0")
